+++ src/main/java/com/ubb/eventapp/model/GroupMemberId.java
@@ -14,8 +14,8 @@ import java.io.Serializable;
 @EqualsAndHashCode
 public class GroupMemberId implements Serializable {
     @Column(name = "id_grupo")
-    private Integer groupId;
+    private Long groupId;
 
     @Column(name = "id_usuario")
-    private Integer userId;
+    private Long userId;
 }
