+++ src/main/java/com/ubb/eventapp/repository/EventReviewRepository.java
@@ -3,5 +3,5 @@ package com.ubb.eventapp.repository;
 import com.ubb.eventapp.model.EventReview;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EventReviewRepository extends JpaRepository<EventReview, Integer> {
+public interface EventReviewRepository extends JpaRepository<EventReview, Long> {
 }
